NDE Update - Work in progress
The Area of code that associated lines based on what delays was
evaluated when attempting to analyse a large industrial program and
found that it was inefficient and needed a (lot) of work.
This block of code is being referred to as the "Decision Engine" and a
new one "NDE - New Decision Engine" is being developed to make adding
new scenarios easier to manage.