Nearly there
This seems to be able to process most of the data, but is getting caught
on a particular section.
We resolve the local alias of a block being processed with the other
data types during calculation of the times, and we then populate the
blockClass object with the raw string.
The problem is that the rawstring still contains the ailias and the
resolved name is lost.  The next step is to extract the Block name from
the S7Statement generated and replace the alias in rawstring with the
resolved block name.
This has been added as comments without comment lines showing an error
to be updated.

Is that light at the end of the tunnel, or a train bombing my way making
things a lot worse?