What's DAMADABOOBOO
Paul:  "DAMADABOOBOO is in your hair".
Me: "What's DAMADABOOBOO"
Paul: "Nothing Yogi"...
If you don't watch Yogi and Booboo, you probably won't get it. But there
is something DAMADA.
The software I have can identify each memory type from the Address
provided.  For local memory it will build a list of variables and the
associated types and cross reference this list as it encounters the
local memory.
the downside is that this does not provide any accessability to Function
Block access or UDTs.
UDTs, I can manage as these will always be at the top of the program
Export and SFBs and SFCs, I can manage as they are predefined prior to
compiling the program, however, I'm currently unsure of how to fix the
program I have to accomodate user compiled function blocks without
rebuilding the entire program... again.

More detail will be provided in an issue in later and a proposed
solition, in the meantime, i can estimate any local accesses by assuming
integer access???