Completed Expedited analysis
Completed Expedited Analysis
The system will now analyse and assign a time to all instructions in a
large industrial program.
This is achieved by reading UDT / device access and assigning a
temporary data type to it, "p".
This is treated as a reasonably worst case reference until a routine is
developed to read through the UDTs, DBs ect. and assign a data type at
runtime.