Work in Progress - Symbol Table Processing
Work in Progress
Adding an addtional function to load a symbol table in SDB Format.
This is to be used to determine where SFBs Etc. have been used and how
to determine the typical length from the Product Literature.
Also gives an option for any other blocks non encountered- be it a
warning or a user management window.