Work in Progress. Still work in progress. The system is processing the OC Program but when pasting the AWL Source or Loading an AWL File from the fileChooser, the process is taking ~10m. The next stage is to remove the inbuilt loops from the main stlAnalyserWindow and move them to a SwingWorker thread. This should operate faster and update a progress window to see what's happening. The Class has been created and methods created, the logic simply needs to be updated.
in this version, the key fix done was to add an additional check to sourceEntry to remove double comments.
After the original comments are removed, the string is checked for additional comments (but not comments only) in lines.
The line is then split into an array by using `String.Split("\\")` and only the first item used.