Update
Funnily enough, I didn't need to place the string read into a separate
string.
The delay was due to the building of the String for reading.
string += string creates a new immutable string every time in the
compiler.  This is why reading to an arraylist was so much faster than a
string, which was required for displaying in the text array.
the end solution was to use a `StringBuilder` object, build the string
using the `append` function and once complete, use the `toString()`
method to populate the text area.