Milestone complete
The software is now successfully marking a complete industrial program
with the marker flag , marker functions and end scan routine.
This may need some tweeking but currently works.
The system needed an additional Check box to allow processing (Go
button) for delay analysis or for marking.
For Delay analysis, any local blocks need to be replaced with the Block
type...
`CALL #THIS_BLOCKMOV` translates to `CALL SFC20`
This way when calculating the block delays, we know the block being
called is SFC20 and need no other information.
However, if the block is compiled into a FB ro SFB and the Call is to
the block `CALL #PARABLOCK` is translated to `CALL FB23`, which will not
compile as there is no instance data block or parametric access defined.
The check box 'process for marking' is used to select between these 2
options and is always visible, not only during debug.