Work in Progress
The size of the marker DB should reflect the number of possible Jumps in
a scan.
This is a user defined value.
This has been added to the main dialog but is still work in progress as
the `predefinedSourceGenerator` is not currently processing MARKER2.AWL
correctly and the endScan is not configured to copy the correct block
size.