Work in progress
Update when attempting to mark another industrial program.
When a block is generated in SCL, it is compiled to STL when downloaded.
This STL can be uploaded into a blank program however the source often
won't compile.
This is causing some issues.
These blocks at present have to be skipped.
(Maybe if I can delete the SCL Blocks from the program before compiling
them?)
the main thing right now is that if you do not add an SDF File, you can
still mark the block and it will compile ok.