Corrected Segment Execution Time
Correct Segment Timer issue.
realised that the Calls are not being identified within the function...
May be ok.
Mark --------- CALL ------- is the same execution time as
Mark ---------------- CALL
so that should even out.
The System is not populating all the Db with a continuous loop.  May
have exceeded some limit here.  this will have to be reflected in
Marker.AWL