Work in progress
Developing a system that will allow translating of a program into a map
using the Jung2 library.
1 - Have the library working and know what to do to create the map.
2 - Have the marking program marking blocks with the exception of calls.
I'm unsure how to deal with this TBH.  if I treat all blocks as single
entities then there will be multiple re-entries and this may mess up
programming analysis as there will no way to determine which is the
correct output.
I'm probably going to have to generate 2 maps.  one where each block
call is treated as a single instance and another where the block call is
a singular instance amongst all calls.