major updates for Usage
Updated the program for better usage outside netbeans.
[ ]    Included a new Dialog to show the segments (Allows for copying
from the results pane as well)
[ ]    Removed the "Process for Marking" check box - all blocks are now
process for marking as the execution time for SFB/SFCs called locally
are no longer required.  This will be determined by Segment details.
[ ]    Updated `marker.AWL` for correct working marker block that auto
rolls.
[ ]    Updated `marker.AWL`  and the main dialog to remove the end scan
block as this is no longer required with the new marker routine.
[ ]    Updated the file chooser logic to select the same path every time
rather than have to navigate to the same path from the default.
[ ]    Updated `marker.AWL` as well as the `markBlockSource()` method to
update the use of loop counters to make resizing the marker DB easier.
Relabelled the JTextField to reflect the new repurposing
[ ]    General fixes.